Add files for some estimates from the 2018 season
These will eventually be deleted, but for now will serve as a record of how the previous Rmd structure looked prior to alterations in this project.

These files are not exectuable and they are not intended to be. I'll go here to copy over new code into building revised Rmds and for testing out the estimation-only code.

I included one example for a set-only opener and one that was both drift and set