Create a function to estimate catch per trip
This function now handles most of what estimate_harvest() used to: calculate average trip-level quantities. Now it handles the data suitability checks (it didn't before) when these averages are calculated. estimate_harvest()'s only job now is to multiply the output of the new function by effort -- it will likely be deleted in the future.