Use <<- instead of = to define some key variables
This was something that took me a couple hours to track down. It has to do with environments and lexical scoping -- still sort of fuzzy on it. But boot_out was defined in one place, and report() was needing to call it when buried deeply in other functions and it couldn't find boot_out. Oddly, this only occurred when using rmarkdown::render() from within the new add-in (see next commit). Defining it as a global variable fixes this problem without having to rewrite all other functions that use report()